Refactor layout and navigation for improved UX
Updated `MainLayout.razor` to introduce a top navigation bar with a gradient background, replacing the sidebar. Restructured the main content area for better layout and visibility of the error UI.

Modified `MainLayout.razor.css` to set the page height to 100vh and enhance scrolling behavior. Removed sidebar styles and applied new styles to the top navigation bar.

Significant changes in `NavMenu.razor`, including the removal of `IDisposable` and a streamlined navigation link structure for improved clarity and accessibility. Simplified the `AuthorizeView` component.

Updated `NavMenu.razor.css` to reflect the new design with enhanced styles for the navbar toggler, link styles, and hover effects, resulting in a modernized appearance.